Rename file, add check to error reading, fix condition around nav item
