+++ frontend/awx/analytics/Reports/Reports.tsx
@@ -49,7 +49,9 @@ export default function Reports() {
         // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
         .json()
         // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
-        .then((r: { error: { keyword: string } }) => setSpecificError(r.error.keyword));
+        .then((r: { error?: { keyword?: string } }) =>
+          setSpecificError(r?.error?.keyword || 'unknown')
+        );
     }
   }, [error]);
