+++ frontend/awx/analytics/Reports/TemplatesTable/ExpandedRowContents.tsx
